Fix the problem where authors were being set incorrectly on historical messages.
When a historic message was *pre-pended*, we were looking at the most recent message's author to decide whether to mark the historic message's author as hidden.

In fact what we need to do is always mark the author of the historic message we're pre-pending as visible; and then depending on whether the oldest existing message's author differs from the historic message's author, we need to mark the oldest existing message's author as hidden.

This is a very rough fix at the moment - it's ripe for refactoring and there are no tests. I can't help feeling that this would all be a bit easier if we introduced explicit javascript models & views for a message and a collection of messages i.e. instead of using the DOM as a model.