Correctly show to the user when Teamstream has been disconnected from PissWhistle.
We determined that when the underlying network connection dies, Chrome does not detect this for the WebSocket. We believe that Chrome should set the WebSocket 'readyState' to 2 (CLOSED) and fire the 'onclose' event (as Safari and Webkit nightly do). We have logged an issue with the Chromium project to determine if this is a bug or expected behaviour.

http://code.google.com/p/chromium/issues/detail?id=76358

So with this in mind we have chosen to use the heartbeat approach to determine the state of the connection. We have removed anything to do with reconnecting to simplify the code.

While investigating the WebSocket behaviour we noticed that the PissWhistle server correctly detects that a client has disconnected, albeit about 60 seconds after the client has disappeared.