Latching into the disconnected state and never checking the connection again was too strict.
By setting up a new timeout in each call to check_connection (whatever the state of the connection), we avoid the latching behaviour.

The return value from PissWhistle.check_connection() seemed to be the opposite of what I would expect and since it doesn't appear to be used anywhere, I've removed it.