Stop getting 'disconnected' messages when we aren't actually disconnected.
After poking around, it looks like the lag between heartbeats is hovering at around 8-12 seconds. I found this by poking around in the Chrome console:

  (new Date().getTime()/1000) - PissWhistle.connection.latestHeartbeat

Running that repeatedly showed that the time was greater than the acceptable lag.

I think this could be because of differences in the time between the server (which sends the heartbeat) and the client (which compares that time to its local time), but I'm not sure of that.

Regardless, even with the disconnected banner visible I could still reliably post messages.